Теперь каждый класс в отдельном файле
