+++ Kippy.Backend/src/Kippy.Domain/Volunteer/DonationRequisite.cs
@@ -1,0 +1,17 @@
+namespace Kippy.Domain.Volunteer;
+
+public sealed class DonationRequisite 
+{
+    private string Name { get; }
+
+    private string Description { get; }
+
+    private string Value { get; }
+
+    public DonationRequisite(string name, string description, string value)
+    {
+        Name = name;
+        Description = description;
+        Value = value;
+    }
+}
