+++ Kippy.Backend/src/Kippy.Domain/Volunteer/FullName.cs
@@ -1,0 +1,17 @@
+namespace Kippy.Domain.Volunteer;
+
+public sealed class FullName
+{
+    public string FirstName { get; }
+    
+    public string LastName { get; }
+    
+    public string MiddleName { get; }
+    
+    public FullName(string lastName, string firstName, string middleName)
+    {
+        LastName = lastName;
+        FirstName = firstName;
+        MiddleName = middleName;
+    }
+}
