+++ Kippy.Backend/src/Kippy.Domain/Volunteer/Pet.cs
@@ -120,40 +120,6 @@ public class Pet : Entity
     }
 }
 
-public sealed class Address 
-{
-    private string City { get; }
-    
-    private string Street { get; }
-    
-    private string Building  { get; }
-    
-    private string Apartment { get; }
-
-    public Address(string city, string street, string building, string apartment)
-    {
-        City = city;
-        Street = street;
-        Building = building;
-        Apartment = apartment;
-    }
-    
-}
-
-public sealed class SpeciesInfo
-{
-    public int SpeciesId { get; }
-    
-    public int BreedId { get; }
-
-    public SpeciesInfo(int speciesId, int breedId)
-    {
-        SpeciesId = speciesId;
-        BreedId = breedId;
-    }
-}
-
-
 public enum HelpStatus
 {
     NeedsHelp,    // Нуждается в помощи
