+++ Kippy.Backend/src/Kippy.Domain/Volunteer/SocialNetwork.cs
@@ -1,0 +1,14 @@
+namespace Kippy.Domain.Volunteer;
+
+public sealed class SocialNetwork
+{
+    public string Name { get; }
+    
+    public string Url { get; }
+
+    public SocialNetwork(string name, string url)
+    {
+        Name = name;
+        Url = url;
+    }
+}
