+++ Kippy.Backend/src/Kippy.Domain/Volunteer/Volunteer.cs
@@ -73,48 +73,3 @@ public class Volunteer : Entity
         );
     }
 }
-
-public sealed class FullName
-{
-    public string FirstName { get; }
-    
-    public string LastName { get; }
-    
-    public string MiddleName { get; }
-    
-    public FullName(string lastName, string firstName, string middleName)
-    {
-        LastName = lastName;
-        FirstName = firstName;
-        MiddleName = middleName;
-    }
-}
-
-public sealed class SocialNetwork
-{
-    public string Name { get; }
-    
-    public string Url { get; }
-
-    public SocialNetwork(string name, string url)
-    {
-        Name = name;
-        Url = url;
-    }
-}
-
-public sealed class DonationRequisite 
-{
-    private string Name { get; }
-
-    private string Description { get; }
-
-    private string Value { get; }
-
-    public DonationRequisite(string name, string description, string value)
-    {
-        Name = name;
-        Description = description;
-        Value = value;
-    }
-}
